// src/modules/user/messages/messages.enum.ts
export enum UserMessage {
  Found = 'حساب کاربری پیدا شد',
  NotFound = 'حساب کاربری پیدا نشد',
  Deleted = 'حساب کاربری حذف گردید',
  Updated = 'حساب کاربری به روز رسانی شد',
  ConflictPhone = 'این شماره موبایل متعلق به حساب کاربری دیگری است',
  ConflictUsername = 'این نام کاربری قبلا انتخاب شده است. نام دیگری انتخاب کنید',
  ConflictEmail = 'این پست الکترونیک متعلق به حساب کاربری دیگری است',
}